Refactor class names in Console Renderer
