+++ src/Sender/Console/Renderer/Binary.php
@@ -6,18 +6,18 @@ namespace Buggregator\Trap\Sender\Console\Renderer;
 
 use Buggregator\Trap\Proto\Frame;
 use Buggregator\Trap\ProtoType;
-use Buggregator\Trap\Sender\Console\RendererInterface;
+use Buggregator\Trap\Sender\Console\Renderer;
 use Buggregator\Trap\Sender\Console\Support\Common;
 use Buggregator\Trap\Sender\Console\Support\Files;
 use Fiber;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * @implements RendererInterface<Frame\Binary>
+ * @implements Renderer<Frame\Binary>
  *
  * @internal
  */
-final class Binary implements RendererInterface
+final class Binary implements Renderer
 {
     private const BYTE_REPLACES = [
         '[ ]' => ' ',
