+++ src/Sender/Console/Renderer/SentryEnvelope.php
@@ -8,16 +8,16 @@ use Buggregator\Trap\Proto\Frame;
 use Buggregator\Trap\ProtoType;
 use Buggregator\Trap\Sender\Console\Renderer\Sentry\Exceptions;
 use Buggregator\Trap\Sender\Console\Renderer\Sentry\Header;
-use Buggregator\Trap\Sender\Console\RendererInterface;
+use Buggregator\Trap\Sender\Console\Renderer;
 use Buggregator\Trap\Sender\Console\Support\Common;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * @implements RendererInterface<Frame\Sentry\SentryEnvelope>
+ * @implements Renderer<Frame\Sentry\SentryEnvelope>
  *
  * @internal
  */
-final class SentryEnvelope implements RendererInterface
+final class SentryEnvelope implements Renderer
 {
     public function isSupport(Frame $frame): bool
     {
