+++ src/Sender/FrameHandler.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Buggregator\Trap\Sender\Console;
+namespace Buggregator\Trap\Sender;
 
 use Buggregator\Trap\Proto\Frame;
 
 /**
  * @internal
  */
-interface HandlerInterface
+interface FrameHandler
 {
     public function handle(Frame $frame): void;
 }
